修改我的页面布局

+++ app/src/main/java/com/wzy/mhealth/activities/TaocanBuyActivity.java
@@ -126,7 +126,7 @@ public class TaocanBuyActivity extends Activity implements Handler.Callback,
                 switch (view.getId()) {
                     case R.id.stepperCustom:
                         number=value;
-                        tv_price.setText(Integer.parseInt(price) * value + "元");
+                        tv_price.setText(Double.parseDouble(price) * value + "元");
                         break;
                 }
             }
